requirements.txt
metagpt>=0.8.0
pytest-playwright==0.4.4